# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login/Register"
                ],
                "summary": "Logowanie",
                "description": "Logowanie odbywa się za pomocą: email, password",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zwrotka Logowania Wygenerowany token należy wkleić do Authorize"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Login/Register"
                ],
                "summary": "Rejestracja",
                "description": "Logowanie odbywa się za pomocą: email, password",
                "operationId": "store",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "user1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zwrotka rejestracji"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "AFTER LOGGED IN"
                ],
                "summary": "Wyloguj się",
                "responses": {
                    "200": {
                        "description": "Zwrotka wylogowania"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/user_info": {
            "get": {
                "tags": [
                    "AFTER LOGGED IN"
                ],
                "summary": "Pobierz dane użytkownika",
                "responses": {
                    "200": {
                        "description": "Zwrotka nowego tokenu"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "get": {
                "tags": [
                    "AFTER LOGGED IN"
                ],
                "summary": "Odświerz token - Po odświerzeniu wprowadź go do aplikacji ponownie",
                "responses": {
                    "200": {
                        "description": "Zwrotka nowego tokenu, należy ją wprowadzić do authorize ponownie"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}